Add files via upload
The motivation behind the color game that I have designed is to brush up or enhance the exposure of different colors to a web developer .Because colors make or break a site.
Choosing a right color palette for a site can exponentially rise the chances of web page's success. This game is fun and informative and gives more exposure to different colors which can be used in real time web dev projects.